Renamed 'Sinigle' to single in package
