+++ src/main/java/frc/robot/RobotContainer.java
@@ -32,8 +32,8 @@ import frc.robot.subsystems.shooter.ColorSensor;
 import frc.robot.subsystems.shooter.Feeder;
 import frc.robot.subsystems.shooter.Flywheel;
 import frc.robot.subsystems.shooter.Intake;
-import frc.robot.subsystems.sinigle_motor.SingleMotorIoReplay;
-import frc.robot.subsystems.sinigle_motor.SingleMotorIoSparkMax;
+import frc.robot.subsystems.single_motor.SingleMotorIoReplay;
+import frc.robot.subsystems.single_motor.SingleMotorIoSparkMax;
 import frc.robot.subsystems.swerve_module.SwerveModuleIoReplay;
 import frc.robot.subsystems.swerve_module.SwerveModuleIoSim;
 import frc.robot.subsystems.swerve_module.SwerveModuleIoTalonFx;
